refactor: api 훅은 api 폴더로 이동 및 한 페이지에서만 사용하는 컴포넌트는 페이지 폴더로 이동하여 폴더 구조 정리
