+++ src/api/hooks/useProductsSummary.ts
