+++ src/pages/main/components/FightingBox.tsx
