+++ src/pages/main/components/GiftItem.tsx
