+++ src/pages/main/components/GiftsRanking.tsx
