+++ src/pages/main/components/SelectFriendBox.tsx
