+++ src/pages/order/OrderPage.tsx
@@ -4,10 +4,10 @@ import { ROUTE_PATH } from "@/routes/paths";
 import { cards } from "@/data/card";
 import type { Card } from "@/types/card";
 import styled from "@emotion/styled";
-import CardSection from "@/components/order/CardSection";
-import SenderSection from "@/components/order/SenderSection";
-import ReceiverSection from "@/components/order/ReceiverSection";
-import GiftInformationSection from "@/components/order/GiftInformationSection";
+import CardSection from "@/pages/order/components/CardSection";
+import SenderSection from "@/pages/order/components/SenderSection";
+import ReceiverSection from "@/pages/order/components/ReceiverSection";
+import GiftInformationSection from "@/pages/order/components/GiftInformationSection";
 import { useUserInfo } from "@/contexts/UserInfoContext";
 import { FormProvider, useForm } from "react-hook-form";
 import type { SubmitHandler } from "react-hook-form";
@@ -15,7 +15,7 @@ import type { OrderFormValue } from "@/types/order";
 import type { OrderRequest } from "@/types/order";
 import { postOrder } from "@/api/order";
 import withUser from "@/hoc/withUser";
-import useProductsSummary from "@/hooks/api/useProductsSummary";
+import useProductsSummary from "@/api/hooks/useProductsSummary";
 import withSuspenseBoundary from "@/hoc/withSuspenseBoundary";
 import { wrapper } from "@/utils/wrapper";
