+++ src/pages/products/ProductPage.tsx
@@ -1,13 +1,13 @@
-import ProductDescriptionSection from "@/components/products/ProductDescriptionSection";
-import ProductDetailSection from "@/components/products/ProductDetailSection";
-import ProductInfoSection from "@/components/products/ProductInfoSection";
-import ProductOrderSection from "@/components/products/ProductOrderSection";
-import ProductReviewSection from "@/components/products/ProductReviewSection";
+import ProductDescriptionSection from "@/pages/products/components/ProductDescriptionSection";
+import ProductDetailSection from "@/pages/products/components/ProductDetailSection";
+import ProductInfoSection from "@/pages/products/components/ProductInfoSection";
+import ProductOrderSection from "@/pages/products/components/ProductOrderSection";
+import ProductReviewSection from "@/pages/products/components/ProductReviewSection";
 import { TABS } from "@/constants/productTab";
 import type { Tab } from "@/constants/productTab";
 import withSuspenseBoundary from "@/hoc/withSuspenseBoundary";
 import withUser from "@/hoc/withUser";
-import useProductsQueries from "@/hooks/api/useProductsQueries";
+import useProductsQueries from "@/api/hooks/useProductsQueries";
 import { wrapper } from "@/utils/wrapper";
 import styled from "@emotion/styled";
 import { useState } from "react";
