+++ src/pages/products/components/ContentItem.tsx
