+++ src/pages/products/components/ProductDescriptionSection.tsx
