+++ src/pages/products/components/ProductDetailSection.tsx
