+++ src/pages/products/components/ProductInfoSection.tsx
