+++ src/routes/router.tsx
@@ -1,12 +1,12 @@
 import { Routes, Route, Navigate } from "react-router";
 import { ROUTE_PATH } from "./paths";
-import MainPage from "@/pages/MainPage";
-import LoginPage from "@/pages/LoginPage";
-import NotFoundPage from "@/pages/NotFoundPage";
-import MyPage from "@/pages/MyPage";
-import OrderPage from "@/pages/OrderPage";
-import ThemesPage from "@/pages/ThemesPage";
-import ProductPage from "@/pages/ProductPage";
+import MainPage from "@/pages/main/MainPage";
+import LoginPage from "@/pages/login/LoginPage";
+import NotFoundPage from "@/pages/not-found/NotFoundPage";
+import MyPage from "@/pages/my/MyPage";
+import OrderPage from "@/pages/order/OrderPage";
+import ThemesPage from "@/pages/themes/ThemesPage";
+import ProductPage from "@/pages/products/ProductPage";
 
 const Router = () => {
   return (
